+++ test/integration/tests/abrmd_policysecret.sh
@@ -35,7 +35,7 @@ tpm2_changeauth -c o ownerauth
 
 # Create Policy
 tpm2_startauthsession -S $session_ctx
-tpm2_policysecret -S $session_ctx -c $TPM_RH_OWNER -o $o_policy_digest ownerauth
+tpm2_policysecret -S $session_ctx -c $TPM_RH_OWNER -L $o_policy_digest ownerauth
 tpm2_flushcontext -S $session_ctx
 rm $session_ctx
 
@@ -47,7 +47,7 @@ tpm2_load -C $primary_ctx -u $seal_key_pub -r $seal_key_priv -o $seal_key_ctx
 
 # Satisfy policy and unseal data
 tpm2_startauthsession --policy-session -S $session_ctx
-echo -n "ownerauth" | tpm2_policysecret -S $session_ctx -c $TPM_RH_OWNER -o $o_policy_digest file:-
+echo -n "ownerauth" | tpm2_policysecret -S $session_ctx -c $TPM_RH_OWNER -L $o_policy_digest file:-
 unsealed=`tpm2_unseal -p"session:$session_ctx" -c $seal_key_ctx`
 tpm2_flushcontext -S $session_ctx
 rm $session_ctx
@@ -63,7 +63,7 @@ unsealed=""
 tpm2_changeauth -c o -p ownerauth
 
 tpm2_startauthsession --policy-session -S $session_ctx
-tpm2_policysecret -S $session_ctx -c $TPM_RH_OWNER -o $o_policy_digest
+tpm2_policysecret -S $session_ctx -c $TPM_RH_OWNER -L $o_policy_digest
 unsealed=`tpm2_unseal -p"session:$session_ctx" -c $seal_key_ctx`
 tpm2_flushcontext -S $session_ctx
 rm $session_ctx